fix `getBinVar()` method
If no subspace flag is passed, the function would get the X category (LOW, SIGNAL, or HIGH) from global, but the order in which the outputs were given was incorrect. `xbin` would be assigned the supspace flag string and `c` would be assigned the integer corresponding to the xbin. 

Swapping `xbin` and `substring` flags fixes this.